backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.119.0",
    "groq>=0.33.0",
    "langchain>=0.3.27",
    "langgraph>=0.6.10",
    "psycopg2>=2.9.11",
    "python-dotenv>=1.1.1",
    "sqlmodel>=0.0.27",
]
